simplify check for when columns load
Previous code required each column to emit an event when they were positioned and owning script had to keep track of when each was completed.

Now we can just wait for the container wrapping the columns to report that everything inside of it has been sorted and resized. Will probably be useful in the future when we can add/subtract columns.