backpack-column-interaction - partial mouse snap
Had an interesting idea to use parenting to simplify managing the backpack's position. Unfortunately that couples a few things together in an undesirable way - mostly problems with draw order + column sorting.

Godot draws nodes in the order of the scene tree, so reparenting the backpack node beneath a column's anchor point makes it stuck behind the other columns in draw order.

A way to work around this is to adjust the order of nodes in the scene tree - bring the currently selected column to the front and it (and its backpack child) will get drawn in front of everything else!

Unfortunately node ordering serves another function in Godot's UI containers (like our Columns node). If node order changes, so does their horizontal position (as determined by columns sort order). Leaves us in a catch-22. Can't leave column order as-is or the backpack will sometimes be drawn behind other columns, but can't change column order or the columns will switch places left-to-right.

The answer is to just not use reparenting as a positioning solution, I think. Unfortunate, though, because the parenting pattern does a good job of describing the relationship between a backpack and its column.