backpack-column-interaction - multiple backpack support
Big feature is how you can drop a backpack into a column to swap places with the other. Also has a moderate amount of event rework and refactor that moved more responsibility to Gameplay.

Gameplay is growing in complexity faster than I'd like, but the backpack-to-backpack interaction really needed something above both of them to moderate interactions (which, for now, is basically just Gameplay). We can always break this backpack-managing code out to a separate class if we want to & use composition to give it the same capabilities.

Need to do a bit of cleanup and hopefully work out a solution to show something in the spot the backpack left behind - a semi-transparent shadow would be great.