backpack-column-interaction - add lerp to pack movement
Code became more fragile with this one, probably. Didn't realize it when I set out to do this feature, but this made the draw order problem rear its head again - now that non-selected packs will move across the screen, they're in danger of getting drawn behind other columns.

Ended up landing on a solution - backpacks will be drawn in front as long as they're in the middle of lerp movement. I think this will look pleasing and intuitive and make it easier to track how backpacks are being exchanged with one another.