backpack-column-interaction - added shadow after pickup
Should be the last functional piece needed in this branch. This faded shadow of the bag you picked up should make it easier to tell where it came from, and the lerp movement added in previous commits helps visually track where the replaced backpack went.

Generally follows the same reparenting strategy used in a couple other places, and the fact it's not visible all the time means I got to cut corners on cleaning it up afterward.