PR Feedback - Apply Remaining Types (#5)
* backpack-column-interaction - implement keyboard-based snap

lots of temporary code in this one, but wanted to prove out an example of shifting a backpack between column positions.

Bit of reorganization here too, Backpacks no longer have any reference to GameplayColumns. For now I think it makes sense to keep columns aware of their backpack, but making this relationship one-way helps keep Backpack's code from getting unnecessarily messy.

* backpack-column-interaction - partial mouse snap

Had an interesting idea to use parenting to simplify managing the backpack's position. Unfortunately that couples a few things together in an undesirable way - mostly problems with draw order + column sorting.

Godot draws nodes in the order of the scene tree, so reparenting the backpack node beneath a column's anchor point makes it stuck behind the other columns in draw order.

A way to work around this is to adjust the order of nodes in the scene tree - bring the currently selected column to the front and it (and its backpack child) will get drawn in front of everything else!

Unfortunately node ordering serves another function in Godot's UI containers (like our Columns node). If node order changes, so does their horizontal position (as determined by columns sort order). Leaves us in a catch-22. Can't leave column order as-is or the backpack will sometimes be drawn behind other columns, but can't change column order or the columns will switch places left-to-right.

The answer is to just not use reparenting as a positioning solution, I think. Unfortunate, though, because the parenting pattern does a good job of describing the relationship between a backpack and its column.

* backpack-column-interaction fix draw order problem

It turned out that there was a simple solution that let us stick to the reparenting strategy - reparent the Backpack to Gameplay while it's picked up to fix front-back draw order without causing trouble for columns.

* backpack-column-interaction - implement snap back

Right now it's leveraging the same methods as moving to a new column. If we want to have some difference in behavior we'll probably need to revisit this implementation in some way.

* backpack-column-interaction - make basic backpack interface

Resource might not be the correct type for Item here. An advantage is we can create and manage different items via Resource files in the editor. Long-term, though, I expect we'll want to just have a big bulk file for defining items for convenience's sake.

* backpack-column-interaction fix temp keyboard control

This wasn't strictly necessary, but it seemed worth keeping around for now. Also helps lay the groundwork for multiple backpacks.

* backpack-column-interaction - multiple backpack support

Big feature is how you can drop a backpack into a column to swap places with the other. Also has a moderate amount of event rework and refactor that moved more responsibility to Gameplay.

Gameplay is growing in complexity faster than I'd like, but the backpack-to-backpack interaction really needed something above both of them to moderate interactions (which, for now, is basically just Gameplay). We can always break this backpack-managing code out to a separate class if we want to & use composition to give it the same capabilities.

Need to do a bit of cleanup and hopefully work out a solution to show something in the spot the backpack left behind - a semi-transparent shadow would be great.

* backpack-column-interaction - remove debug logging

* backpack-column-interaction - add lerp to pack movement

Code became more fragile with this one, probably. Didn't realize it when I set out to do this feature, but this made the draw order problem rear its head again - now that non-selected packs will move across the screen, they're in danger of getting drawn behind other columns.

Ended up landing on a solution - backpacks will be drawn in front as long as they're in the middle of lerp movement. I think this will look pleasing and intuitive and make it easier to track how backpacks are being exchanged with one another.

* backpack-column-interaction - added shadow after pickup

Should be the last functional piece needed in this branch. This faded shadow of the bag you picked up should make it easier to tell where it came from, and the lerp movement added in previous commits helps visually track where the replaced backpack went.

Generally follows the same reparenting strategy used in a couple other places, and the fact it's not visible all the time means I got to cut corners on cleaning it up afterward.

* backpack-column-interaction - docs and reorganization

* De-Indent Section Comments

* Apply Types with Spacing Like Godot Wiki

* Alphabetize Methods for Readability

* Apply Remaining Types for Clarity

---------

Co-authored-by: SamBumgardner <sambumgardner1@gmail.com>