bug-0 - fix double event handling bug.
This happened because we mixed input handling styles. _input() method is called on any kind of input event happening - including things like mouse movement. By using event.is_action_pressed, we determine if the incoming event is of the specific kind of action we're looking for. If the event trigger _input is anything else, it won't pass the conditional check.

What we had in place before was different, though, it was a static method that checked if the 'gameplay_select' action had happened this frame. On the frame we clicked, this conditional evaluated would evaluate true every time it was checked. This means that if we clicked the mouse while moving, it would trigger to calls to _input (one for mouse movement and one for the click) and both would pass the conditional check because the mouse had been clicked during that frame.

This change is small, but it ensures that we only do _handle_backpack_selection for input events that are actually for 'gameplay_select' being triggered. No more having unrelated events occuring in the same frame as the click triggering the backpack behavior.