regions - basic graphic integration
While working through this I realized that the column's header should probably also be part of the ColumnContents scene. Seems like an excellent case for a base scene + inherited scenes for regions and customers. Both need the header graphic and name logic, but the nodes needed to fill out the body of the column will be pretty different.