pack-polish - prevent changing day early
There wasn't a great way to resolve the situation where someone mashes 'next day' button. Partially because they'd be advancing the day without seeing what is going to end up in their bag.

There might be some ideas for how to make it a smoother experience for folks that want to skip everything, but I think prioritizing a bit of polish over speed that ignores the game is a good plan.