pack-polish - art cleanup and handle missing pack.
Need to revisit this in the morning, there's got to be a cleaner way to organize the logical flow of 'next day' happening in a column contents object. Right now it's pretty loosey-goosey and tied to visual bits in tweens. Functional, but going to be more of a headache if we have to touch this again later.